+++ src/store/reduсers/data/data.js
@@ -1,0 +1,14 @@
+import {ActionType} from '../../utils/consts';
+import {extend} from '../../utils/funcs';
+
+
+export const reducer = (state, action) => {
+  switch (action.type) {
+    case ActionType.CHANGE_ACTIVE_MOVIE:
+      return extend(state, {
+        activeMovie: action.payload.activeMovie,
+      });
+    default:
+      return state;
+  }
+};
