+++ src/store/reduсers/name-space.js
@@ -1,0 +1,7 @@
+const NameSpace = {
+  DATA: `DATA`,
+  PLAYER: `PLAYER`,
+  USER: `USER`,
+};
+
+export default NameSpace;
