+++ src/store/reduсers/reducer.js
@@ -1,22 +1,11 @@
-import {ActionType} from '../../utils/consts';
-import {extend} from '../../utils/funcs';
+import {combineReducers} from 'redux';
+import {reducer as data} from './data/data';
+import {reducer as player} from './player/player';
+import {reducer as user} from './user/user';
+import NameSpace from './name-space';
 
-
-export const reducer = (state, action) => {
-  switch (action.type) {
-    case ActionType.CHANGE_ACTIVE_MOVIE:
-      return extend(state, {
-        activeMovie: action.payload.activeMovie,
-      });
-    case ActionType.OPEN_FULL_SCREEN_PLAYER:
-      return extend(state, {
-        isVideoPlayer: true,
-      });
-    case ActionType.CLOSE_FULL_SCREEN_PLAYER:
-      return extend(state, {
-        isVideoPlayer: false,
-      });
-    default:
-      return state;
-  }
-};
+export default combineReducers({
+  [NameSpace.DATA]: data,
+  [NameSpace.PLAYER]: player,
+  [NameSpace.USER]: user,
+});
